Move the class IncompleteCompanyRepresentativeState to namespace Library.States.Companies
