+++ src/Library/HighLevel/Entrepreneurs/Searcher.cs
@@ -22,7 +22,7 @@ namespace Library.HighLevel.Entrepreneurs
         /// </summary>
         /// <param name="publications"></param>
         /// <param name="category"></param>
-        public List<MaterialPublication> SearchByCategory(List<MaterialPublication> publications, MaterialCategory category)
+        public List<MaterialPublication> SearchByCategory(IList<MaterialPublication> publications, MaterialCategory category)
         {
            List<MaterialPublication> searchResultA = new List<MaterialPublication>();
            foreach (var item in publications)
@@ -40,7 +40,7 @@ namespace Library.HighLevel.Entrepreneurs
         /// </summary>
         /// <param name="publications"></param>
         /// <param name="keyword"></param>
-        public List<MaterialPublication> SearchByKeyword(List<MaterialPublication> publications, string keyword)
+        public List<MaterialPublication> SearchByKeyword(IList<MaterialPublication> publications, string keyword)
         {
            List<MaterialPublication> searchResultB = new List<MaterialPublication>();
            foreach (var item in publications)
@@ -59,7 +59,7 @@ namespace Library.HighLevel.Entrepreneurs
         /// <param name="publications"></param>
         /// <param name="locationSpecified"></param>
         /// <param name="distanceSpecified"></param>
-        public List<MaterialPublication> SearchByLocation(List<MaterialPublication> publications, Location locationSpecified, double distanceSpecified)
+        public List<MaterialPublication> SearchByLocation(IList<MaterialPublication> publications, Location locationSpecified, double distanceSpecified)
         {
            List<MaterialPublication> searchResultC = new List<MaterialPublication>();
            foreach (var item in publications)
